create controller for user roles

+++ source/com/qit/server/repositories/UserRoleRepository.java
@@ -3,11 +3,12 @@ package com.qit.server.repositories;
 import com.qit.server.models.SimpleUser;
 import com.qit.server.models.UserRole;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 
 import java.math.BigInteger;
 import java.util.Set;
 
-public interface UserRoleRepository extends JpaRepository<UserRole, BigInteger> {
+public interface UserRoleRepository extends JpaRepository<UserRole, Long> {
 
 //    @Query(value = "SELECT * FROM user_roles WHERE username = ?1", nativeQuery = true)
     Set<UserRole> findRolesByUser(SimpleUser user);
